requirements.txt
camel-ai
siliconflow-deepseek
python-dotenv
jsonschema
